Add files via upload
My group and I were tasked to create a program with team members based on the python skills learned so far. We decided to create a math quiz that teachers or parents can use to test their kids knowledge. The code would ask the user a math question and tell them if the answer is right or wrong. In the end, it would tell the user how many they got correct over how many they got incorrect. We tried to make the code as simple as possible, so we created functions on one file and then ran the code in a different file.